r82xx: eliminate PLL lock check
In practice it almost always locks, and when it doesn't, raising VCO
current fails to help and it still doesn't lock. While it may be nice to
warn of a lock failure or slow lock, this slows down retuning, and
doesn't really help much. Eliminating the read operation speeds up
retuning considerably.